+++ app/controllers/administrate/posts_controller.rb
@@ -1,0 +1,33 @@
+# frozen_string_literal: true
+
+module Administrate
+  class PostsController < ApplicationController
+    def index
+      @posts = Post.all.order("created_at DESC")
+    end
+
+    def new
+      @post = Post.new
+    end
+
+    def create
+      @post = Post.new(post_params)
+
+      if @post.save
+        redirect_to(@post)
+      else
+        render("new")
+      end
+    end
+
+    def show
+      @post = Post.find(params[:id])
+    end
+
+    private
+
+    def post_params
+      params.require(:post).permit(:title, :body)
+    end
+  end
+end
