+++ app/models/post.rb
@@ -1,4 +1,6 @@
 # frozen_string_literal: true
 
 class Post < ApplicationRecord
+  validates :title, presence: true, length: { minimum: 5 }
+  validates :body, presence: true
 end
