+++ app/views/administrate/posts/index.html.erb
@@ -1,0 +1,8 @@
+<% @posts.each do |post| %>
+<div class="post_wrapper">
+  <h2 class="title"><%= link_to post.title, [:administrate, post] %></h2>
+  <p class="body"><%= truncate(post.body, length:100) %></p>
+  <p class="date"><%= time_ago_in_words(post.created_at) %></p>
+</div>
+<% end %>
+<%= link_to  "New Post", new_administrate_post_path %>
