+++ app/views/administrate/posts/new.html.erb
@@ -1,0 +1,27 @@
+<div id="page_wrapper">
+<h1>Make Something</h1>
+
+<%= form_for :post, url: administrate_posts_path do |f| %>
+<% if @post.errors.any? %>
+<div id="errors">
+  <h2><%= pluralize(@post.errors.count, "error") %>Stopped this post from saving</h2>
+  <ul>
+    <% @post.errors.full_messages.each do |msg| %>
+    <li><%= msg %></li>
+    <% end %>
+  </ul>
+</div>
+<% end %>
+<p>
+  <%= f.label :title %><br>
+  <%= f.text_field :title %>
+</p>
+<p>
+  <%= f.label :body %><br>
+  <%= f.text_area :body %>
+</p>
+<p>
+  <%= f.submit %>
+</p>
+<% end %>
+</div>
