+++ app/views/administrate/posts/show.html.erb
@@ -1,0 +1,13 @@
+<div id="post_content">
+  <h1 clas="title">
+    <%= @post.title %>
+  </h1>
+
+  <p class="date">
+    Submitted <%= time_ago_in_words(@post.created_at) %> ago
+  </p>
+
+  <p class="body">
+  <%= @post.body %>
+  </p>
+</div>
