+++ app/views/layouts/application.html.erb
@@ -12,8 +12,30 @@
   </head>
 
   <body>
-    <main class="container mx-auto mt-28 px-5 flex">
+    <div id="sidebar">
+      <ul>
+        <li class="category"><%= link_to  "Tumblr", root_path %></li>
+        <li><%= link_to "Posts", root_path%></li>
+        <li>About</li>
+      </ul>
+
+      <p class="sigin_in">User Login</p>
+    </div>
+    <div id="main_content">
+      <div id="header">
+        <p>Post Feed</p>
+        <div class="button">
+          <button class="button"><%= link_to "Make Post", new_administrate_post_path %></button>
+          <button class="button">Log Out</button>
+        </div>
+      </div>
+
+      <% flash.each do |name, msg| %>
+      <%= content_tag(:div, msg, class: "alert") %>
+      <% end %>
       <%= yield %>
+    </div>
+    <main class="container mx-auto mt-28 px-5 flex">
     </main>
   </body>
 </html>
