--- app/views/posts/index.html.erb
@@ -1,6 +1,0 @@
-<% @posts.each do |post| %>
-<div class="post_wrapper">
-  <h2 class="title"><%= link_to post.title, post %></h2>
-  <p class="date"><%= time_ago_in_words(post.created_at) %></p>
-</div>
-<% end %>
