--- app/views/posts/new.html.erb
@@ -1,15 +1,0 @@
-<h1>Make Something</h1>
-
-<%= form_for :post, url: posts_path do |f| %>
-<p>
-  <%= f.label :title %><br>
-  <%= f.text_field :title %>
-</p>
-<p>
-  <%= f.label :body %><br>
-  <%= f.text_area :body %>
-</p>
-<p>
-  <%= f.submit %>
-</p>
-<% end %>
