--- app/views/posts/show.html.erb
@@ -1,11 +1,0 @@
-<h1 clas="title">
-  <%= @post.title %>
-</h1>
-
-<p class="date">
-  Submitted <%= time_ago_in_words(@post.created_at) %> ago
-</p>
-
-<p class="body">
-<%= @post.body %>
-</p>
